refactor: use updatePanelOrder as a camera method
