+++ src/camera/Camera.ts
@@ -10,7 +10,8 @@ import Panel from "../core/panel/Panel";
 import AnchorPoint from "../core/AnchorPoint";
 import * as ERROR from "../const/error";
 import { ALIGN, CIRCULAR_FALLBACK, DIRECTION, EVENTS, ORDER } from "../const/external";
-import { checkExistence, find, getFlickingAttached, getProgress, includes, parseAlign, toArray } from "../utils";
+import { checkExistence, find, getFlickingAttached, getProgress, getStyle, includes, parseAlign, toArray } from "../utils";
+import { ValueOf } from "../type/internal";
 
 import { CameraMode, BoundCameraMode, CircularCameraMode, LinearCameraMode } from "./mode";
 
@@ -40,6 +41,7 @@ class Camera {
   private _visiblePanels: Panel[];
   private _anchors: AnchorPoint[];
   private _needPanelTriggered: { prev: boolean; next: boolean };
+  private _panelOrder: ValueOf<typeof ORDER>;
 
   // Internal states getter
   /**
@@ -217,6 +219,14 @@ class Camera {
     }
   }
 
+  /**
+   * {@link https://developer.mozilla.org/en-US/docs/Web/CSS/direction direction} CSS property applied to the camera element(`.flicking-camera`)
+   * @ko 카메라 엘리먼트(`.flicking-camera`)에 적용된 {@link https://developer.mozilla.org/en-US/docs/Web/CSS/direction direction} CSS 속성
+   * @type {string}
+   * @readonly
+   */
+  public get panelOrder() { return this._panelOrder; }
+
   // Options Getter
   /**
    * A value indicating where the {@link Camera#alignPosition alignPosition} should be located at inside the viewport element
@@ -257,6 +267,7 @@ class Camera {
     this._checkTranslateSupport();
 
     this._updateMode();
+    this.updatePanelOrder();
 
     return this;
   }
@@ -556,8 +567,30 @@ class Camera {
     const actualPosition = this._position - this._alignPos - this._offset + this._circularOffset;
 
     el.style[this._transform] = flicking.horizontal
-      ? `translate(${flicking.panelOrder === ORDER.RTL ? actualPosition : -actualPosition}px)`
+      ? `translate(${this._panelOrder === ORDER.RTL ? actualPosition : -actualPosition}px)`
       : `translate(0, ${-actualPosition}px)`;
+
+    return this;
+  }
+
+  /**
+   * Update direction to match the {@link https://developer.mozilla.org/en-US/docs/Web/CSS/direction direction} CSS property applied to the camera element
+   * @ko 카메라 엘리먼트에 적용된 {@link https://developer.mozilla.org/en-US/docs/Web/CSS/direction direction} CSS 속성에 맞게 방향을 업데이트합니다
+   * @return {this}
+   */
+  public updatePanelOrder(): this {
+    const el = this._el;
+    const flicking = getFlickingAttached(this._flicking);
+    if (flicking.horizontal) {
+      const direction = getStyle(el).direction;
+      if (direction !== this._panelOrder) {
+        this._panelOrder = direction === ORDER.RTL ? ORDER.RTL : ORDER.LTR;
+        if (flicking.initialized) {
+          flicking.control.controller.updateDirection();
+          this.applyTransform();
+        }
+      }
+    }
 
     return this;
   }
