+++ test/unit/Flicking.spec.ts
@@ -96,37 +96,6 @@ describe("Flicking", () => {
         expect(flicking.currentPanel).to.equal(flicking.camera.anchorPoints[0].panel);
       });
     });
-
-    it("should guarantee behavior for panels placed in RTL order", async () => {
-      const flicking = await createFlicking(
-        El.viewport().setDirection("rtl").add(
-          El.camera()
-            .add(El.panel("300px"))
-            .add(El.panel("300px"))
-            .add(El.panel("300px"))
-        )
-      );
-
-      await simulate(flicking.element, { deltaX: 1000 });
-
-      expect(flicking.index).to.equal(2);
-    });
-
-    it("should update direction of panels when resize occurs", async () => {
-      const viewportEl = El.viewport().setDirection("ltr").add(
-        El.camera()
-          .add(El.panel("300px"))
-          .add(El.panel("300px"))
-          .add(El.panel("300px"))
-      );
-      const flicking = await createFlicking(viewportEl);
-
-      viewportEl.setDirection("rtl");
-      await flicking.resize();
-      await simulate(flicking.element, { deltaX: 1000 });
-
-      expect(flicking.index).to.equal(2);
-    });
   });
 
   describe("Properties", () => {
